Workspace SaveAs Guids replacement
